CoCo3 .DSK disk image of Joust progress
CoCo3 .DSK disk image file showing the JOUST code running the Highscores screen and the big JOUST logo attract screen.  The third screen is a demo screen, which normal shows some basic game play.  I modified the code so it draws all the platforms to see how they will look and how much of the game will fit on the CoCo 3 screen.  It looks like only a few rows along the bottom will be missing, other than that it looks good.
To use the .DSK image on a 512k CoCo 3 just type LOADM"GO and hit enter.
I have a video of this on youtube but the video doesn't do the colours justice.